BinaryReader: Add size guard when reading bytes.
In case of requesting really large buffers (often by accident), we can
end up requesting up to 2GB of memory for a single byte array and
immediately running out of heap space, usually crashing the app.

If more than 128kB, read in chunks of 4k and siphon through an
expandable byte buffer. In case we run out of bytes to read, we're saved
by the actual input stream...