Progress: Show updating tasks at the bottom.
Move completed tasks out of the list of updating tasks when the updating
list is more than double of the max tasks.

First try to keep current order by moving completed task be be "before"
current tasks. If this is not enough, sort all completed to the top, and
move those ouf of the "current" tasks.