Resolve FileWatcher directory, not file.
If the watched file's parent directory is resolved to canonical path,
but not the file itself we get more consistency to what we listen to:
change to the file we point to as seen from the directory we specify.

This means that given:

```
/path/link -> /other
/path/file -> /other/file
/other/file
```

If you listen to /path/link/file, you get notified to changed to the
file content.

If you listen to /path/file, you get notified of changes to the symlink
that points to /other/file, in case it points to a different file, but
not to changes to the content of that file.

PRIOR:

All files were resolved to canonical file paths, so only changes to the
content of the actual underlying file would be detected. Also the file
path would always be the full canonical path of the resolved file when
called back to the watchers.